requirements.txt
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.65.0
python-dotenv>=1.0.0
